+++ apps/staking/pages/index.tsx
@@ -13,7 +13,7 @@ import {
 } from "evmos-wallet";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 
-import { TermOfServices, Container } from "ui-helpers";
+import { TermOfServices, Container, MavaWidget } from "ui-helpers";
 import { MixpanelProvider } from "tracker";
 function SnackbarsInternal() {
   const valueRedux = useSelector((state: StoreType) => getAllSnackbars(state));
@@ -48,6 +48,7 @@ export default function Home() {
                   <>
                     <SnackbarsInternal />
                     <StatefulHeader />
+                    <MavaWidget />
                     <div className="container mx-auto mb-auto overflow-auto">
                       <Content />
                     </div>